Added str to all enum classes so that all variables of enum can map to str
